pyproject.toml
[project]
name = "destep"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jaydebeapi>=1.2.3",
    "jpype1>=1.6.0",
    "loguru>=0.7.3",
    "pypinyin>=0.55.0",
    "sqlalchemy>=2.0.45",
    "typer>=0.21.1",
]

[tool.ruff]
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
    "SIM",
    "N",
]

ignore = [
    "E501", # Line too long
    "N805"
]

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
